Introduce Persistence module
