+++ hydra-node/hydra-node.cabal
@@ -289,6 +289,7 @@ test-suite tests
     Hydra.NodeSpec
     Hydra.OptionsSpec
     Hydra.PartySpec
+    Hydra.Persistence
     Hydra.SnapshotStrategySpec
     Paths_hydra_node
     Spec
@@ -361,5 +362,5 @@ test-suite tests
     , websockets
     , yaml
 
-  build-tool-depends: hspec-discover:hspec-discover -any
+  build-tool-depends: hspec-discover:hspec-discover
   ghc-options:        -threaded -rtsopts
